Improvements to the sysmon-mqtt install script
The systemd unit is now always "reinstalled"; makes for a smoother (if
somewhat lengthier) upgrade process.

Also, the systemd unit now aims to wait until the network is up before
starting sysmon-mqtt; this should prevent the script from not starting
properly at boot (when it would be called before the network was up).