Several tweaks to androidtv-container
The "create_user" script failed after the initial run of the container
(as the requested user was already created). This in turn caused the
adb server not to start and the entire container to fail... The script
is now much more resilient.

Furthermore, pin the version of adb to install (in the hopes of
preventing unwanted surprises when building the container at a later
point in time).