Automate waiting for RPi CM4 to be in-stock
Instead of keeping an eye on https://rpilocator.com/, use Home Assistant
as RSS-reader...

This required a small (generally useful) update to the persistent
notification script: It's now possible to provide a URI for "Action 1".
In this case, the notification action button opens the provided URI
(URL, but also app intents, etc.) on the phone.