Goldplate the doorbell announcement
As the announcement text gets rendered when the announcement is queued,
it's not possible to have dynamic elements with regards to the moment of
playback (e.g. "5 minutes ago") in there.

Introduce some "magic markers" (e.g., %TIME_AGO%) that are processed
just before the announcement gets played. This way it becomes possible
to add simply dynamisms to the announcements...