Add Single-board computers "subview" to Network
Includes a detailed overview of the single-board computers on the
network, including an indication of whether updates are availabe and
which devices need a restart (after automatically installing updates).

The most important devices are still listed on the main Network-view.
Their widgets now also include a (clearer) indication of their update
and "restart required" status.