Expand "TTS queue" and "TTS announce"
With the introduction of Piper (local TTS) it became a bit less
attractive to build very long TTS announcements (as the time it takes
to generate them makes the responses feel less responsive). The solution
is to cut up the responses in smaller pieces and queue them one after
the other (with the delay now spread out in between the individual parts
instead of fully front loaded).

"TTS announce" Already did the queuing, so that was the easy part.
Expanded "TTS queue" to add some related functionality: There are now
two announcement channels, the original "attention"-channel which allows
for less time-sensitive announcements to be queued and the (basically
identical) "announce"-channel which allows for more immediate
announcements to be queued up too.
This allows for announcements to be generated by different automations
and queued either for immediate release ("tts_announce") or for later
release ("tts_attention").

Furthermore, added support for "announcing" samples (e.g., MP3 files).
This way all "announcements" can be looped through a single queue (that
of the "TTS announce"-script), thus preventing one from overwriting the
other – everything is now neatly queued and played FIFO.