Make Hourly "Dong!" use "TTS announce"
This way there's no more possibility for it to conflict with other
announcements running at the top of the hour.

Use the new "Everyone asleep" delay timer to determine of we're going to
bed (instead of inferring it from the state of the lights). Use the
wakeup snooze time timer to prevent interfering with the wakeup alarm.

Finally, do the full "Dong!" only at noon, for all other hours use
something a bit less "disruptive".