Refactor out *all* old template definitions...
...and remove all unnecessary customisations from "customize.yaml"
(replacing them with proper templates and sensor entities with a unique
ID wherever possible). The only thing left are customisations that
cannot be made in any other way (at the moment).

All (non-template and non-MQTT) sensor definitions are now in the
"sensors/" and "sensors_binary/"-folders, all (binary_)sensor template
definitions are in the "templates/"-folder, and all (manually
configured) MQTT sensors are in the "mqtt/"-folder.

Removed a handful of unused tadoº template sensors (exposing attributes
that weren't used anywhere) and merged the "Time of Day (numeric)"
sensor as an attribute into the main "Time of Day" sensor (purely used
for graphing on in Lovelace, so no need for a separate sensor).