Failed troubleshooting of Attic hallway lights
In an attempt to further troubleshoot the lingering issues with the
Zigbee relay (`lumi.relay.c2acn01`) in the Attic hallway, I managed to
make matters rather worse than better...

Managed to recover without having to climb on a ladder and physically
reset the relay, but it did result in a renaming of all entities
involved (to more descriptive names, so didn't bother renaming them
back to their old ids). In the process it appears I did manage to break
all metering functionalities; anyway, they were not used.

For now everything appears to be working as well as previously, at least
from a "turning on the lights" perspective. I've ordered a couple
(potential) alternative Zigbee relays, so most likely to be continued...