Attic Hallway lights plot twist 🤯
No relevant code change, but something that might give me a good chuckle
in the future: There is nothing inherently wrong with the firmware of
the Zigbee relay powering the Bathroom fan / the Attic hallway lights.

Apparently the chimney cover is leaking and (with past few weeks'
wonderful "Summer" weather) it got quite wet at the bottom of the
chimney pipe _where the relay is located_...

The relay simply got wet and started to bug out like crazy (a bit scary
when you think of it). Quite a coincidence that I noticed: It was
raining heavily outside while I was listing for the relay to *click*,
and instead I heard dripping noises 🤐

For now the relay is hanging outside the pipe (on the "dry" Bathroom
side) until the "Bouwvak" is over and I can get someone to put a new
hood on the chimney. Will then replace it with an undamaged one _and_
put it in an electrical box (which I of course should have done to begin
with).