Expand Garden irrigation functionality
Instead of relying on a (backstopped) manual start/stop, use two input
helpers to set the desired watering volumes and run the irrigation via
a script that stops automatically when the desired volume is reached
(and decrements the input helpers towards zero as it does so).

This way, only a manual start is required and the watering volume can
be adjusted "real-time" (ie, while watering) as it becomes clear there's
either too little or too much water going into the garden...

In the future the watering volumes could be automatically (re)set (every
night) taking into account environmental conditions (as a prelude
towards full automation).

Also greatly enhanced the irrigation dashboard: It now includes an
overview of net evapotranspiration and soil moisture over the past 10
days to aid in better judging if and how much watering is required.