Tweak the empty/disabled integrations
They appear to be throwing up errors regarding "config_flow" not being
available when (I guess, relevant) bluetooth devices are detected.

Perhaps it's somehow reading the original configuration and then
merging the custom one on top of it? Let's see if disabling both the
"config_flow" and providing an empty set of Bluetooth identifiers
resolves the issue...