Reorganise the Home Alarm automations
Too many of them were clustered together in "alert.yaml" and
"away.yaml". Made a more sensible split between the various triggering
automations (causing the alarm to sound) and notifying automations
(which "inform" about the alarm state through mobile notifications, TTS,
and the state of the lights in the house).

The automation that silences the Home Alarm from the Bedroom after it
gets triggered was moved to a separate file. It was extended to make
unintentionally silencing the alarm less likely. Furthermore, it is now
"TTS aware" and releases a set of message indicating why the alarm is
sounding (these are set alongside the regular mobile notifications).

Not too much functionality was added, but due to moving a lot of stuff
around it's quite a big commit. Hasn't been fully tested yet, so some
more tweaks might follow...