Add a calamity for extremely high PM levels
The CO₂ sensors also detect PM10/2.5; extremely high PM levels might
indicate a fire in/near the house and/or some other kind of calamity
for which a warning is warrented.

Such a warning is now provided. During the day it'll send out mobile
notifications and announce via TTS. At night the actual sirens are used
(as the situation might be actually dangerous).

To properly cope with this, a setup similar to silencing the Home
Alarm from the Bedroom is now also in place for the calamities (the
smoke detector and leak sensors were already triggering the sirens).
This includes queueing up TTS announcements to be released when the
sirens are silenced.

Again not fully tested yet, so expect some further tweaks in the future.