Fix Dockerfile linting errors
While I was at it, also set up hadolint properly (now that an ARM64
binary is available). This uncovered some (minor) issues in two
Dockerfiles, which leads to the point that nothing is ever easy:

Had to include a local copy of just-containers' public key (it's not
available on https://keybase.io anymore).
Also, when installing wget with "--no-install-recommends", the
ca-certificates are not installed (which leads to issues trying to
download stuff over HTTPS 🤐).