Include ShellCheck and hadolint in the linter
And learned something about Bash process substitution along the way...

Especially for the femon-script it seems a more appropriate approach (as
the loop now runs _while_ sampling instead of after all samples have
been collected; quicker and more memory efficient).

For the linter less so, but the ShellCheck example (SC2046) explicitly
mentions a linter as a use-case in its example.