Improve "everyone_asleep_delay" implementation
Instead of running out unconditionally (and then checking if the lights
are still off), short-circuit (and reset) the timer when the lights are
switched back on.

This way the timer can now (more reliably) be used to determine if we're
in the process of going to bed – ie, as intended (instead of having to
observe the state of the lights as well).