Use named pipe instead of temporary file
Bit of gold-plating, but nonetheless interesting to figure out how this
works: Instead of writing the round-trip times to a temporary file and
reading them back from there, open a named pipe, attach it to file
descriptor 3, and use that to communicate between the main process and
the forked off "ping" process...

Also, always echo "sysmon-mqtt <version>" when the script starts. The
"--version" argument simply makes it exit immediately after doing so.