Replace custom linting with pre-commit
Mainly for its convenience in generalising linter configuration; not so
much for its use as package manager. Most of the linters in-use need to
be installed locally anyway for use with VS Code. So, there's little
benefit in pulling in a lot of external stuff (and then having to deal
with other people's idiosyncrasies where it comes to linter setup).

The main advantage (not yet in this commit) is that the "Lint codebase"
GitHub Action can now use the exact same approach as the pre-commit
hook (and it can easily be tested locally).

With regards to that, several small tweaks to linter configurations were
made to accommodate the current codebase.