Pin pre-commit and linter versions
Not doing so will probably come around and bite me in the future...

It was a nice idea to store versions as repository variables, but that
made the setup-script _and_ caching logic unnecessarily complicated and
prevents "point-in-time" reproducibility/traceability.
So, for now a "low-tech" approach by just hardcoding versions across
README.md, the setup-script and pre-commit configuration.

In the process, moved from using the APT version of ShellCheck (which
for Ubuntu 22.04 is still 0.8.0) to simply pulling the latest version
(0.9.0) from GitHub. The process is similar to that for both hadolint
(which is not available via Ubuntu's APT) and shfmt (which is also
a bit outdated in APT), so there's little extra cost. It does –
additionally – confer the same "point-in-time" benefits mentioned above.

There was also a nasty oversight in README.md (hadolint uses "x86_64"
instead of "amd64", so "arch=$(dpkg --print-architecture)" would've
never worked).

Finally, while I was at it, added caching to the sysmon-mqtt Docker
workflow too (already "natively" available in the Docker action).