Hue scene heuristic tweaks
Observed some odd behaviour in the Kitchen. Difficult to diagnose as the
"raw"-heuristic sensors are excluded from the recorder; include them in
the recorder for now.

Also, in preparation for potential future fixes, split of the Kitchen
light motion automations to a separate file (slowly splitting up the
overly large lighting automation configurations).