Tweaks to the Hue scene heuristic
Ran into an issue where the heuristic failed as a result of the Kitchen
in "Ambiance" with the Cupboards in "Activity" being identical to the
Kitchen in "Daytime". Specifically:
  { num_lights: 14, brightness: 77, xy_color[0.46, 0.35] }

For now the offending rule was removed from "Daytime" (I might have made
made a mistake adding it in the first place; let's see...). If that
doesn't cut it, might have to factor in more attributes (e.g.
"color_mode"), or (perhaps better/easier) match on exact combinations of
existing properties (instead mixing and matching).
Although the attributes move around a bit, they do appear to mostly show
up in identical combinations.

Logging has been extended to try and capture the situation if it occurs
again.

The heuristic was tuned to be "cognitively" easier (the implementation
is still complicated): Whenever the Cupboards are in "Activity" and the
Kitchen is _not_ "Bright (custom)", the Kitchen's previous scene is
detected. When the Kitchen is "Bright (custom)" the Cupboards are
detected as <Kitchen> (instead of "Activity"). This way there's less
ambiguity in the detection rules.