Split Ensuite off from Living Room
Replaced a faulty "dumb" Zigbee bulb with a proper Hue bulb in the
Living Room; took this is an opportunity to properly manage the Ensuite
as a separate room (as it's generally controlled separately from the
rest of the Living Room).

Cleaned up the Hue scene heuristic for the Living Room (with the goal
of keeping brightness- and xy_color-settings grouped together instead of
allowing them to be freely mixed-and-matched).

Finally, major gold-plating to nicely show the fireplace in the list
of Hue scenes (not really a scene, but fits there better than as a
button of its own that never actually needs to be pressed).