Correct a devious typo...
The goal of the "[^A-Za-z0-9_ -.]"-pattern was to include all alpha-
numeric characters, the underscore, space, hyphen, and dot. But, how it
was specified, the last part (" -.") was interpreted as the range of
characters from space to dot... 🤐

This caused some unwanted characters (e.g., the parentheses) to get
through. They were caught by the following "tr", but did cause multiple
consecutive underscores to appear in the cleaned name.