Coax NPM into behaving properly...
Somehow trying to run markdownlint and/or Prettier "dynamically" via npx
doesn't work reliably in GitHub Actions... 🤨 Worked most of the time;
failed occasionally – at which point rerunning the job used to resolve
the issue. Except for today, at which point it seems to have completely
broken down 😢

So, give up on trying to run them dynamically via npx and simply install
the packages (globally) as part of the install-script. Probably the most
consistent approach anyway (a similar one is already used for the pip-
packages). As an additional benefit, all version pinning now happens
in the install-script, so the GitHub Action cache only has to rely on
that file and not on the pre-commit configuration itself anymore.