Replace (overly complex) Hue scene heuristic
In hindsight not a good idea – in my defence: The button events and Hue
scene entities the new approach relies on were not available when I
initially conceived the heuristic...

Although it worked, the heuristic detection was rather unstable. Both
from a heuristic perspective (the values it relied on were not the same
from one day to the next), and from a maintainability perspective.

The new approach relies on button presses and scene changes to keep
track of the active Hue scene. This is _a lot_ easier and _a lot_ more
stable.

An additional benefit is that it allows the Hue scene settings to be
changed without having to dive into Home Assistant and figure out the
new values for the heuristic.
Changing/adding buttons and scenes does require Home Assistant updates,
but that happens a lot less often.

Additionally, scene chances directly in the Hue app aren't reflected
any more in Home Assistant. Since I actually only use the Hue app to
configure the lights, this has turned out to be a benefit instead of
(the expected) drawback: It's much easier (and more convenient) to
experiment with lighting settings now...