Upgrade to Home Assistant 2023.12
Including various custom components and JavaScript modules.

Removed the "Google DNS (ping)"-sensor. HA 2023.12 doesn't allow its
interval to be configured anymore (and the default 5-minute one is a bit
too long for my taste). There's a proper workaround available, but
sysmon-mqtt provides a drop-in replacement...

As the sysmon-mqtt sensor runs outside of Home Assistant, it also
appears to be less impacted by the load on HA (which seems to have
caused spikes in RTT not caused by network delay).