Remove WaterRower RESTful-command
Worked "okay", but it seems a recent update to my Garmin broke the
rowing app on the watch (or at the very least, no more data was coming
in while the connection was properly up).

As a result, I installed the WaterRower app on my phone and must
conclude that works substantially better (and my Garmin still tracks the
most relevant statistics without the ANT+ sensor connected).

So, that idea went out the window. The Pi0 is now the new Guestroom
speaker (to which this automation now broadcasts).