Implement "Auto-arm when away"
Prevents the Home Alarm from automatically arming when there's no-one
home. A better/easier implementation of the "Guest-mode" idea: Don't
automatically arm the alarm when there's guests in the house and (in a
subsequent commit) make several automations that used to be dependent on
whether "group.family" was home, instead depend on the state of the Home
Alarm...

Also, use a Tile card instead of the Alarm Panel card to interact with
the Home Alarm.