Replace "group.family" with Home Alarm state
In most cases, it's better to look at the Home Alarm state to determine
whethere someone is home. It covers all kinds of situations that were
previously complex (ie, guests in the house; phone ran out of battery)
at the cost of a slight increase in verbosity.

Generally speaking, the check is either "disarmed" (someone is in the
house and awake), or _not_ "armed_away"/"armed_vacation" (someone is
present in the house).

Not all cases have been replaced (as sometimes it's useful to explicitly
check whether _we_ are at home). The "everyone_asleep" boolean is still
required in many places (as there's no guarantee the alarm is armed at
night, for example when guests are over it sometimes is not).

All of this comes at the risk of scaring the living daylights out of a
burglar (as the house might start to talk to them); there's no great
benefit in attempt to guard against that.
Furthermore, in a situation where the alarm is sounding, announcements
can now also be generated. Generally speaking those would be high
urgency (e.g. "The house is on fire") so perhaps for the better that
they now also come through.