Use Bluetooth for Qinping Air Monitor Lite
Seems to be working just as well as the custom Xiaomi Miot component,
but has the benefit of being "native" (and less complex for just a
single set of sensors).

For now just updated the dashboards so I can confirm the sensors work
reliably. Once that's done, the automations will follow.