Clean-up the Lovelace "Network"-view
Switched from my trusty DD-WRT units to Unifi, so all that stuff became
non-functional. Again, the Unifi-app provides a proper overview, so no
sense in burdening Home Assistant with that.

Only thing that is a bit of a shame as that I can't map bandwidth usage
and RTT over a long period of time anymore...