Use HA Voice satellites for announcements
More streamlined approach; saves having the manually implement the
"chime" part of the announcement. Also creared a separate script to
play samples (instead of having that glued on top of the TTS-script).

Overall a cleaner and simpler approach. Moved the Piper TTS over from an
RPi4 to an RPi5, so there's no more need to split long messages up into
smaller parts to get a quicker TTS-response.