Fix/tweak A(a)rlo capture implementation
With the recently updated configuration, the front camera is capturing
more reliably again. Arlo-connectivity is a bit spotty though, so
triggering snapshots on it causes a lot of unnecessary snapshots to be
created. Simplest solution was to simply disable that trigger (main
point anyway is/was to refresh the snapshot every two-hours).

Run all snapshot/capture automations in parallel; also wait for Arlo
connectivity for the daily snapshots (and increase the timeout to
ten-minutes); remove the pause-functionality (never really used, removed
it from the dashboard last year already).