Don't announce when detecting motion outside
Primarily triggered on the neighbour's cat (which started to get greatly
annoying). Furthermore, the current state of the Aarlo-integration (as a
result of the subpar service offered by Arlo itself) is such that
there's no (timely) way to detect what caused the motion. The
"last"-motion sensors regularly take multiple hours to update... 😵