Replace ADB with Plex and Android TV Remote
The ADB integration seems to be running on its last legs. Also, I'm
throwing in the towel with regards to trying to keep the Android TV from
going to sleep (and becoming unavailable). Don't know what happened, but
as of late even the dedicated wakelock-app running on the device seems
unable to keep it awake. I guess in the future we can't have nice things
anymore.

As such, the ADB integration offers no more benefit over the Plex
integration. Setting up and maintaining it is actually way more complex
than using Plex (and it doesn't appear actively maintained on the Home
Assistant side), so it goes out the window 👋
In combination with the Android TV Remote (easiest way to tell if the
device is powered on, _and_ it knows which app is running) all existing
automations can be retooled without loss of functionality.

In the future I might tool up template-entity to more easily capture the
full state of the television, but this is good enough for now...