Massive overhaul of the lighting automations
Had (over the years) automated myself somewhat into a corner... Mostly
cleaned up now (not fully tested yet though 😇). Split over rooms
instead of "times of day". Cleanly splits the automations; better
manageability at the cost of some more duplication.

The primary daily flow for most rooms is now as follows:
- At wakeup time, the lights go on. Either to a "reduced" setting (based
  on sun relative to house) or directly into daytime-lighting (guarded
  by a lux threshold)
- During the day, daytime-lighting manages the lights (and also toggles
  between the "reduced" and regular setting based on sun relative to
  house)
- At dusk, lights are switched into evening mode (and from there further
  managed)

The amount of exceptions and transitions is greatly reduced, so the
daily flow is now mostly straightforward, with only a handful of
affordances made to cope with the distinction between the "reduced", and
regular daytime-lighting states.